Basic UI Coding done
Finished coding the basic UI stuff, it is pretty messy but the engine is
hard to work with based on my experience so far. At least when it comes
to HUD positioning but then again I may have gone overboard trying to
work it out all out for any resolutions.