Threading, vsync and other cool things
Threaded the map generation so its now way faster than what it used to
be.

Added vsync support to the engine for Windows as the native GLFW
implementation was not actually working.

Optimised the way the time display works. Added background actors that
are active if the application is resizeable or fullscreen.