NPC Phase
Begun work on NPCs, converted map into a singleton for easier access, as
the containers in it are used by a lot of different objects.

Moreover, added support to change window name for Windows builds as well
as a special function to append notices on window names as well as
remove them.

Also, other changes to game logic so that the world timer starts paused
and required the user to press Space to unpause it.

And lastly... added purses! Inventory system now is a combination of
InventoryBag objects for storing resources and also InventoryPurse,
usually one but that can easily be changed and extended, for storing
gold along with the usual assortment of helper functions and checks so
that there can be gold limits and such placed on potential player
actors. These checks can be ignored by passing a boolean in the
functions as FALSE instead of the default TRUE.

Currently, not plans to extend the purse functionality beyond basics
because it would require time that could otherwise be spend on the
actual honours project (unfortunately most of the current work is only
the precursor to the honours project work).